+++ src/Game.css
@@ -1,0 +1,8 @@
+.gameOver {
+    min-height: 25px;
+    padding: 5px;
+}
+
+.playAgain {
+    cursor: pointer;
+}
